// src/app/core/users/users.types.ts
export interface Users
{
    groupId: number
     nameGroup: string
     colorGroup: string
     userId: number
     firstName: string
     lastName: string
     email: string
     urlImage: string
     sedeId: number
     sedeName: string
     roule: string
     disable: number
     disableDate?: string
     hidden: number
     currentGroupId: number
}